refactor(viewer): single-source the per-event-type render switch
EventStream and TrackingTimeline each carried a 34-case switch over the
same RecordedEvent union: EventStream's EventDetail (full detail row) and
TrackingTimeline's eventSummary (compact tooltip line). Adding an event
type or changing a field meant editing both in lock-step, and they had
already drifted (e.g. truncation lengths, terminalCommand duration).

Move both switches verbatim into one module, src/utils/eventDisplay.tsx,
exporting eventDetail(event) and eventSummary(event, sessionStartTime).
Each component imports and calls its function; the bodies are unchanged.
EventStream's call site goes from <EventDetail event={e}/> to
{eventDetail(e)} (the old EventDetail was a plain function component with
no hooks/state, so output and reconciliation are identical). Trim each
component's format imports to the helpers it still uses.

Now the event-type union is handled in exactly one place.
