+++ recording-viewer/src/components/EventStream.tsx
@@ -2,7 +2,8 @@ import { useState, useMemo, useRef, useEffect, useCallback } from 'react';
 import { Virtuoso, type VirtuosoHandle } from 'react-virtuoso';
 import type { Annotation, RecordedEvent, EventType } from '../types.ts';
 import { ALL_LABELS } from '../types.ts';
-import { formatOffset, formatDuration, shortenUri, formatDebugSessionMeta, formatBreakpointLocation } from '../utils/format.ts';
+import { formatOffset } from '../utils/format.ts';
+import { eventDetail } from '../utils/eventDisplay.tsx';
 import { EventBadge } from './EventBadge.tsx';
 
 interface Props {
@@ -60,248 +61,6 @@ function stripAnsi(text: string): string {
         .replace(/\r(?!\n)/g, '')
         .trim();
     /* eslint-enable no-control-regex */
-}
-
-export function EventDetail({ event }: { event: RecordedEvent }) {
-    switch (event.type) {
-        case 'eqSnapshot':
-            return (
-                <span className="event-detail">
-                    EQ: <strong>{Math.round(event.eq * 100)}%</strong>
-                    <span className={`confidence-tag ${event.confidence}`}>{event.confidence}</span>
-                </span>
-            );
-        case 'buildResult':
-            return (
-                <span className="event-detail">
-                    {event.buildFailed ? 'BUILD FAILED' : event.successful ? 'PASSED' : `${event.errorCount} error(s)`}
-                    {event.failedTests.length > 0 && ` | ${event.failedTests.length} test(s) failed`}
-                </span>
-            );
-        case 'textChange': {
-            const MAX_INLINE = 16;
-            let insertedAll = '';
-            let deletedTotal = 0;
-            for (const c of event.changes) {
-                insertedAll += c.text;
-                deletedTotal += c.rangeLength;
-            }
-            const hasInsert = insertedAll.length > 0;
-            const hasDelete = deletedTotal > 0;
-
-            const inlineText = insertedAll.length <= MAX_INLINE
-                ? <code className="inline-text">{insertedAll}</code>
-                : <><code className="inline-text">{insertedAll.slice(0, MAX_INLINE)}</code> +{insertedAll.length - MAX_INLINE} chars</>;
-
-            let op: React.ReactNode;
-            if (hasInsert && hasDelete) {
-                op = <span className="change-preview">replaced {deletedTotal} &rarr; {inlineText}</span>;
-            } else if (hasInsert) {
-                op = <span className="change-preview">inserted {inlineText}</span>;
-            } else {
-                op = <span className="change-preview">deleted {deletedTotal} chars</span>;
-            }
-
-            return (
-                <span className="event-detail">
-                    {shortenUri(event.uri)} | {op}
-                </span>
-            );
-        }
-        case 'save':
-            return <span className="event-detail">{shortenUri(event.uri)}</span>;
-        case 'diagnostics':
-            return (
-                <span className="event-detail">
-                    {shortenUri(event.uri)} | {event.diagnostics.length} diagnostic(s)
-                </span>
-            );
-        case 'fileSwitch':
-            return (
-                <span className="event-detail">
-                    {shortenUri(event.fromUri)} &rarr; {shortenUri(event.toUri)}
-                </span>
-            );
-        case 'sessionStart':
-            return (
-                <span className="event-detail">
-                    Exercise {event.exerciseId}
-                    {event.participantId && ` | ${event.participantId}`}
-                </span>
-            );
-        case 'sessionEnd':
-            return <span className="event-detail">Exercise {event.exerciseId}</span>;
-        case 'irisChatMessage':
-            return (
-                <span className="event-detail">
-                    {event.direction === 'sent' ? 'SENT' : 'RECV'}:&nbsp;
-                    {event.content.length > 80 ? event.content.slice(0, 80) + '...' : event.content}
-                    {event.messageId && <span className="event-meta"> id:{event.messageId}</span>}
-                </span>
-            );
-        case 'irisChatSendAttempt':
-            return (
-                <span className="event-detail">
-                    {event.status.toUpperCase()}:&nbsp;
-                    {event.content.length > 80 ? event.content.slice(0, 80) + '...' : event.content}
-                    {event.errorMessage && <span className="event-error"> — {event.errorMessage}</span>}
-                </span>
-            );
-        case 'irisChatFeedback':
-            return (
-                <span className="event-detail">
-                    msg:{event.messageId} | {event.helpful ? 'helpful' : 'not helpful'}
-                </span>
-            );
-        case 'windowFocus':
-            return <span className="event-detail">{event.focused ? 'focused' : 'blurred'}</span>;
-        case 'fileSnapshot':
-            return <span className="event-detail">{shortenUri(event.uri)}</span>;
-        case 'selectionChange':
-            return (
-                <span className="event-detail">
-                    {shortenUri(event.uri)} | L{event.selections[0]?.startLine ?? 0}:{event.selections[0]?.startCharacter ?? 0}
-                    {event.kind && ` (${event.kind})`}
-                </span>
-            );
-        case 'visibleRangeChange':
-            return (
-                <span className="event-detail">
-                    {shortenUri(event.uri)} | L{event.visibleRanges[0]?.startLine ?? 0}-L{event.visibleRanges[0]?.endLine ?? 0}
-                </span>
-            );
-        case 'intervention':
-            return (
-                <span className="event-detail">
-                    {event.action.toUpperCase()} | {event.level}
-                    {' '} EQ: <strong>{Math.round(event.eq * 100)}%</strong>
-                    {event.triggerType && ` | ${event.triggerType}`}
-                </span>
-            );
-        case 'eqEngineState':
-            return (
-                <span className="event-detail">
-                    {event.snapshots.length} snapshot(s) | EQ: <strong>{Math.round(event.currentEQ * 100)}%</strong>
-                    <span className={`confidence-tag ${event.confidence}`}>{event.confidence}</span>
-                </span>
-            );
-        case 'viewNavigation':
-            return (
-                <span className="event-detail">
-                    {event.from} &rarr; {event.to}
-                </span>
-            );
-        case 'panelVisibility':
-            return (
-                <span className="event-detail">
-                    {event.panel} | {event.visible ? 'visible' : 'hidden'}
-                </span>
-            );
-        case 'testResultsOverviewView':
-            if (event.action === 'opened') {
-                return (
-                    <span className="event-detail">
-                        Test results overview opened | {event.passedTests}/{event.totalTests} passed ({event.failedTests} failed)
-                    </span>
-                );
-            }
-            return (
-                <span className="event-detail">
-                    Test results overview closed | {formatDuration(event.durationMs)} ({event.closeReason})
-                </span>
-            );
-        case 'taskFeedbackView':
-            if (event.action === 'opened') {
-                return (
-                    <span className="event-detail">
-                        Task "{event.taskName}" opened | {event.passedTests}/{event.totalTests} passed ({event.failedTests} failed)
-                    </span>
-                );
-            }
-            return (
-                <span className="event-detail">
-                    Task "{event.taskName}" closed | {formatDuration(event.durationMs)} ({event.closeReason})
-                </span>
-            );
-        case 'configurationSnapshot':
-            return (
-                <span className="event-detail">
-                    struggleDetection:{event.struggleDetectionEnabled ? 'on' : 'off'} | interventions:{event.showInterventions ? 'on' : 'off'}
-                </span>
-            );
-        case 'configurationChange': {
-            const parts: string[] = [];
-            if (event.changes.struggleDetectionEnabled !== undefined) {
-                parts.push(`struggleDetection:${event.changes.struggleDetectionEnabled ? 'on' : 'off'}`);
-            }
-            if (event.changes.showInterventions !== undefined) {
-                parts.push(`interventions:${event.changes.showInterventions ? 'on' : 'off'}`);
-            }
-            return <span className="event-detail">{parts.join(' | ')}</span>;
-        }
-        case 'terminalCommand': {
-            const exitOk = event.exitCode === 0;
-            return (
-                <span className="event-detail terminal-detail">
-                    <code>{event.command.length > 60 ? event.command.slice(0, 60) + '...' : event.command}</code>
-                    {' '}exit: <strong className={exitOk ? 'terminal-exit-ok' : 'terminal-exit-fail'}>{event.exitCode ?? '?'}</strong>
-                    {' '}({Math.round(event.durationMs / 1000)}s)
-                    {event.outputTruncated && ' [truncated]'}
-                </span>
-            );
-        }
-        case 'terminalOpenClose':
-            return (
-                <span className="event-detail">
-                    {event.action} | {event.terminalName}
-                </span>
-            );
-        case 'fileSnapshotError':
-            return (
-                <span className="event-detail">
-                    {shortenUri(event.uri)} | {event.reason}
-                </span>
-            );
-        case 'fileCreate':
-            return <span className="event-detail">{shortenUri(event.uri)}</span>;
-        case 'fileDelete':
-            return <span className="event-detail">{shortenUri(event.uri)}</span>;
-        case 'fileRename':
-            return (
-                <span className="event-detail">
-                    {shortenUri(event.oldUri)} &rarr; {shortenUri(event.newUri)}
-                </span>
-            );
-        case 'textDocumentOpen':
-            return <span className="event-detail">{shortenUri(event.uri)}</span>;
-        case 'textDocumentClose':
-            return <span className="event-detail">{shortenUri(event.uri)}</span>;
-        case 'debugSession':
-            return (
-                <span className="event-detail">
-                    {event.action}{formatDebugSessionMeta(event.sessionName, event.sessionType)}
-                </span>
-            );
-        case 'breakpointChange': {
-            const first = event.breakpoints[0];
-            return (
-                <span className="event-detail">
-                    {event.action} | {event.breakpoints.length} breakpoint(s)
-                    {first && ` | ${formatBreakpointLocation(first.uri, first.line)}`}
-                </span>
-            );
-        }
-        case 'submission':
-            return (
-                <span className="event-detail">
-                    SUBMIT {event.status.toUpperCase()} | participation {event.participationId}
-                    {event.commitMessage && ` | "${event.commitMessage.length > 60 ? event.commitMessage.slice(0, 60) + '...' : event.commitMessage}"`}
-                    {event.failureReason && <span className="event-error"> — {event.failureReason}</span>}
-                </span>
-            );
-        default:
-            return null;
-    }
 }
 
 function InlineAnnotationInput({ onSubmit, onCancel }: {
@@ -541,7 +300,7 @@ export function EventStream({ events, sessionStartTime, annotations, enabledType
                         {formatOffset(event.timestamp - sessionStartTime)}
                     </span>
                     <EventBadge type={event.type} />
-                    <EventDetail event={event} />
+                    {eventDetail(event)}
                     {isTermCmd && (
                         <span className="expand-hint">{isTermExpanded ? '▾' : '▸'}</span>
                     )}
