+++ recording-viewer/test/debugRendering.test.tsx
@@ -1,8 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { render } from '@testing-library/react';
 import type { RecordedEvent } from '../src/types';
-import { eventSummary } from '../src/components/TrackingTimeline';
-import { EventDetail } from '../src/components/EventStream';
+import { eventSummary, eventDetail } from '../src/utils/eventDisplay';
 
 const debugSession: RecordedEvent = {
     type: 'debugSession', timestamp: 1000, action: 'started',
@@ -27,13 +26,13 @@ describe('debug event rendering', () => {
     });
 
     it('EventDetail renders debugSession action + name', () => {
-        const { container } = render(<EventDetail event={debugSession} />);
+        const { container } = render(<>{eventDetail(debugSession)}</>);
         expect(container.textContent).toContain('started');
         expect(container.textContent).toContain('Launch');
     });
 
     it('EventDetail renders breakpointChange with 1-based line', () => {
-        const { container } = render(<EventDetail event={breakpointChange} />);
+        const { container } = render(<>{eventDetail(breakpointChange)}</>);
         expect(container.textContent).toContain('added');
         expect(container.textContent).toContain(':10');
     });
