+++ recording-viewer/test/submissionRendering.test.tsx
@@ -1,8 +1,7 @@
 import { describe, it, expect } from 'vitest';
 import { render } from '@testing-library/react';
 import type { RecordedEvent } from '../src/types';
-import { eventSummary } from '../src/components/TrackingTimeline';
-import { EventDetail } from '../src/components/EventStream';
+import { eventSummary, eventDetail } from '../src/utils/eventDisplay';
 
 const started: RecordedEvent = { type: 'submission', timestamp: 1000, status: 'started', participationId: 42, commitMessage: 'wip' };
 const succeeded: RecordedEvent = { type: 'submission', timestamp: 2000, status: 'succeeded', participationId: 42, exerciseId: 7, commitMessage: 'final' };
@@ -19,12 +18,12 @@ describe('submission event rendering', () => {
     });
 
     it('EventDetail renders the status and participation', () => {
-        const text = render(<EventDetail event={succeeded} />).container.textContent ?? '';
+        const text = render(<>{eventDetail(succeeded)}</>).container.textContent ?? '';
         expect(text).toContain('SUCCEEDED');
         expect(text).toContain('42');
     });
 
     it('EventDetail renders the failure reason', () => {
-        expect((render(<EventDetail event={failed} />).container.textContent ?? '')).toContain('merge-conflict');
+        expect((render(<>{eventDetail(failed)}</>).container.textContent ?? '')).toContain('merge-conflict');
     });
 });
